refactor(backend): move fitness specific models to module
