+++ apps/backend/src/models.rs
@@ -232,3 +232,106 @@ pub mod media {
         pub volumes: Option<i32>,
     }
 }
+
+pub mod fitness {
+    use super::*;
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseForce {
+        Static,
+        Pull,
+        Push,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseLevel {
+        Beginner,
+        Intermediate,
+        Expert,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseMechanic {
+        Isolation,
+        Compound,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseEquipment {
+        #[serde(rename = "medicine ball")]
+        MedicineBall,
+        Dumbbell,
+        #[serde(rename = "body only")]
+        BodyOnly,
+        Bands,
+        Kettlebells,
+        #[serde(rename = "foam roll")]
+        FoamRoll,
+        Cable,
+        Machine,
+        Barbell,
+        #[serde(rename = "exercise ball")]
+        ExerciseBall,
+        #[serde(rename = "e-z curl bar")]
+        EZCurlBar,
+        Other,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseMuscle {
+        Abdominals,
+        Abductors,
+        Adductors,
+        Biceps,
+        Calves,
+        Chest,
+        Forearms,
+        Glutes,
+        Hamstrings,
+        Lats,
+        #[serde(rename = "lower back")]
+        LowerBack,
+        #[serde(rename = "middle back")]
+        MiddleBack,
+        Neck,
+        Quadriceps,
+        Shoulders,
+        Traps,
+        Triceps,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "snake_case")]
+    pub enum ExerciseCategory {
+        Powerlifting,
+        Strength,
+        Stretching,
+        Cardio,
+        #[serde(rename = "olympic weightlifting")]
+        OlympicWeightlifting,
+        Strongman,
+        Plyometrics,
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    #[serde(rename_all = "camelCase")]
+    pub struct Exercise {
+        #[serde(rename = "id")]
+        pub identifier: String,
+        pub name: String,
+        pub force: Option<ExerciseForce>,
+        pub level: ExerciseLevel,
+        pub mechanic: Option<ExerciseMechanic>,
+        pub equipment: Option<ExerciseEquipment>,
+        pub primary_muscles: Vec<ExerciseMuscle>,
+        pub secondary_muscles: Vec<ExerciseMuscle>,
+        pub instructions: Vec<String>,
+        pub category: ExerciseCategory,
+        pub images: Vec<String>,
+    }
+}
